Refactor 3MF materials into subpackages
Split the monolithic import/export material code into dedicated packages for better organization and maintainability. New packages:
- io_mesh_3mf/export_materials (base, textures, pbr, passthrough)
- io_mesh_3mf/import_materials (base, textures, pbr, passthrough)
Also added standalone export_trianglesets.py and import_trianglesets.py, and reworked export_utils.py and import_3mf.py to use the new layout. CHANGELOG.md updated with a Code Organization entry and notes about backward-compatible re-exports to preserve the existing API. This refactor prepares the codebase for easier future enhancements and smaller, focused modules.
