+++ io_mesh_3mf/export_utils.py
@@ -12,52 +12,75 @@
 # <pep8 compliant>
 
 """
-Utility functions for 3MF export: geometry, materials, archive, and metadata handling.
+Core utility functions for 3MF export.
+
+This module contains:
+- Archive management (create_archive, must_preserve, write_core_properties, write_thumbnail)
+- Unit conversion
+- Geometry writing (check_non_manifold, format_transformation, write_vertices, write_triangles)
+- Metadata writing
+
+Materials Extension functionality has been moved to the export_materials package.
+Triangle Sets Extension functionality has been moved to export_trianglesets module.
+
+For backward compatibility, this module re-exports functions from:
+- export_materials.base
+- export_materials.textures
+- export_materials.pbr
+- export_materials.passthrough
+- export_trianglesets
 """
 
 import base64
 import collections
 import datetime
 import itertools
-import json
 import logging
 import os
 import tempfile
 import xml.etree.ElementTree
 import zipfile
-from typing import Optional, Dict, List, Tuple
+from typing import Optional, Dict, List
 
 import bpy
-import bpy_extras.node_shader_utils
 import mathutils
 
 from .annotations import Annotations
 from .constants import (
     MODEL_NAMESPACE,
     MODEL_DEFAULT_UNIT,
-    MATERIAL_NAMESPACE,
     CORE_PROPERTIES_LOCATION,
     CORE_PROPERTIES_NAMESPACE,
     DC_NAMESPACE,
     DCTERMS_NAMESPACE,
-    TRIANGLE_SETS_NAMESPACE,
-    TEXTURE_MIMETYPE_PNG,
-    TEXTURE_MIMETYPE_JPEG,
-    TEXTURE_REL,
-    RELS_NAMESPACE,
     conflicting_mustpreserve_contents,
 )
 from .metadata import Metadata
 from .unit_conversions import blender_to_metre, threemf_to_metre
 
-# Orca Slicer paint_color encoding for filament IDs
-# This matches CONST_FILAMENTS in OrcaSlicer's Model.cpp
-# Index 0 = no color (base extruder), 1-32 = filament IDs
-ORCA_FILAMENT_CODES = [
-    "", "4", "8", "0C", "1C", "2C", "3C", "4C", "5C", "6C", "7C", "8C", "9C", "AC", "BC", "CC", "DC",
-    "EC", "0FC", "1FC", "2FC", "3FC", "4FC", "5FC", "6FC", "7FC", "8FC", "9FC", "AFC", "BFC",
-    "CFC", "DFC", "EFC",
-]
+# Re-export from export_materials for backward compatibility
+from .export_materials import (  # noqa: F401
+    ORCA_FILAMENT_CODES,
+    material_to_hex_color,
+    get_triangle_color,
+    collect_face_colors,
+    write_materials,
+    write_prusa_filament_colors,
+    extract_pbr_from_material as _extract_pbr_from_material,
+    write_pbr_display_properties as _write_pbr_display_properties,
+    detect_textured_materials,
+    detect_pbr_textured_materials,
+    write_textures_to_archive,
+    write_texture_relationships,
+    write_texture_resources,
+    get_or_create_tex2coord,
+    write_pbr_textures_to_archive,
+    write_pbr_texture_display_properties,
+    write_passthrough_materials,
+)
+
+# Re-export from export_trianglesets for backward compatibility
+from .export_trianglesets import write_triangle_sets  # noqa: F401
 
 log = logging.getLogger(__name__)
 
@@ -273,1833 +296,6 @@ def unit_scale(context: bpy.types.Context, global_scale: float) -> float:
     # Scale from Blender scene units to 3MF units.
     scale *= blender_unit_to_metre / threemf_unit_to_metre
     return scale
-
-
-# =============================================================================
-# Material Handling
-# =============================================================================
-
-def material_to_hex_color(material: bpy.types.Material) -> Optional[str]:
-    """
-    Extract hex color string from a Blender material.
-
-    Tries Principled BSDF first (for node-based materials), falls back to diffuse_color.
-    Skips default gray (0.8, 0.8, 0.8) from Principled BSDF.
-    Converts from linear (Blender) to sRGB (3MF hex) color space.
-
-    :param material: The Blender material to extract color from.
-    :return: Hex color string like "#RRGGBB" or None if no material.
-    """
-    if material is None:
-        return None
-
-    color = None
-
-    # Try Principled BSDF first for materials with node setup
-    if material.use_nodes and material.node_tree:
-        principled = bpy_extras.node_shader_utils.PrincipledBSDFWrapper(
-            material, is_readonly=True
-        )
-        base_color = principled.base_color
-        # Check if it's not the default gray (0.8, 0.8, 0.8)
-        if (base_color and not (abs(base_color[0] - 0.8) < 0.01
-                                and abs(base_color[1] - 0.8) < 0.01
-                                and abs(base_color[2] - 0.8) < 0.01)):
-            color = base_color
-
-    # Fall back to diffuse_color for simple materials
-    if color is None:
-        color = material.diffuse_color[:3]
-
-    # Blender's diffuse_color is already in sRGB display space - no conversion needed
-    red = min(255, round(color[0] * 255))
-    green = min(255, round(color[1] * 255))
-    blue = min(255, round(color[2] * 255))
-    return "#%0.2X%0.2X%0.2X" % (red, green, blue)
-
-
-def get_triangle_color(mesh: bpy.types.Mesh, triangle: bpy.types.MeshLoopTriangle,
-                       blender_object: bpy.types.Object) -> Optional[str]:
-    """
-    Get the color for a specific triangle from its face's material assignment.
-
-    :param mesh: The mesh containing the triangle.
-    :param triangle: The triangle to get the color for.
-    :param blender_object: The object the mesh belongs to.
-    :return: Hex color string like "#RRGGBB" or None if no color.
-    """
-    if triangle.material_index < len(blender_object.material_slots):
-        material = blender_object.material_slots[triangle.material_index].material
-        return material_to_hex_color(material)
-    return None
-
-
-def collect_face_colors(blender_objects: List[bpy.types.Object],
-                        use_mesh_modifiers: bool,
-                        safe_report) -> Dict[str, int]:
-    """
-    Collect unique face colors from all objects for Orca color zone export.
-
-    This extracts colors from material assignments per face. Each face can have its own
-    material, allowing solid per-face coloring (perfect for cubes with different colored sides).
-    :param blender_objects: List of Blender objects to extract colors from.
-    :param use_mesh_modifiers: Whether to apply modifiers when getting mesh.
-    :param safe_report: Callable for reporting errors/warnings.
-    :return: Dictionary mapping color hex strings to filament indices (0-based).
-    """
-    unique_colors = set()
-    objects_processed = 0
-
-    for blender_object in blender_objects:
-        if blender_object.type != 'MESH':
-            continue
-
-        objects_processed += 1
-        log.info(f"Processing object: {blender_object.name}")
-
-        # Get evaluated mesh with modifiers applied
-        if use_mesh_modifiers:
-            dependency_graph = bpy.context.evaluated_depsgraph_get()
-            eval_object = blender_object.evaluated_get(dependency_graph)
-        else:
-            eval_object = blender_object
-
-        try:
-            mesh = eval_object.to_mesh()
-        except RuntimeError:
-            log.warning(f"Could not get mesh for object: {blender_object.name}")
-            continue
-
-        if mesh is None:
-            log.warning(f"Mesh is None for object: {blender_object.name}")
-            continue
-
-        # Extract colors from face material assignments
-        log.info(f"Object {blender_object.name}: {len(mesh.vertices)} vertices, {len(mesh.polygons)} faces")
-
-        # Get all materials used by faces
-        materials_used = set()
-        for poly in mesh.polygons:
-            if poly.material_index < len(blender_object.material_slots):
-                materials_used.add(poly.material_index)
-
-        log.info(f"Object uses {len(materials_used)} different materials across its faces")
-
-        # Extract color from each material
-        for mat_idx in materials_used:
-            if mat_idx < len(blender_object.material_slots):
-                material = blender_object.material_slots[mat_idx].material
-                if material is None:
-                    log.warning(f"Material slot {mat_idx} is empty")
-                    continue
-
-                color_hex = material_to_hex_color(material)
-                if color_hex:
-                    unique_colors.add(color_hex)
-                    log.info(f"Face color: {color_hex} from material '{material.name}'")
-
-        # Clean up the temporary mesh
-        eval_object.to_mesh_clear()
-
-    # Sort colors for consistent ordering and create index mapping
-    # IMPORTANT: Start at index 1 because Orca's paint_color codes:
-    #   - "" (empty/no attribute) = no paint, use object base material
-    #   - "4" = filament 1, "8" = filament 2, etc.
-    # So all colored faces need paint_color attributes starting from index 1
-    sorted_colors = sorted(unique_colors)
-    color_to_index = {color: idx + 1 for idx, color in enumerate(sorted_colors)}
-
-    log.info(f"Collected {len(unique_colors)} unique colors from {objects_processed} objects for Orca export")
-    log.info(f"Colors: {sorted_colors}")
-
-    # Report to user
-    if objects_processed == 0:
-        safe_report({'ERROR'}, "No mesh objects found to export!")
-    else:
-        safe_report({'INFO'}, f"Detected {len(unique_colors)} face colors for Orca export: {sorted_colors}")
-
-    return color_to_index
-
-
-def write_materials(resources_element: xml.etree.ElementTree.Element,
-                    blender_objects: List[bpy.types.Object],
-                    use_orca_format: bool,
-                    vertex_colors: Dict[str, int],
-                    next_resource_id: int,
-                    export_pbr: bool = True
-                    ) -> Tuple[Dict[str, int], int, str, Optional[xml.etree.ElementTree.Element]]:
-    """
-    Write the materials on the specified blender objects to a 3MF document.
-
-    Supports:
-    - Core spec <basematerials> with displaycolor
-    - PBR display properties (metallic, specular, translucent workflows)
-    - Orca Slicer colorgroup format
-
-    :param resources_element: A <resources> node from a 3MF document.
-    :param blender_objects: A list of Blender objects that may have materials.
-    :param use_orca_format: Whether to use Orca Slicer format.
-    :param vertex_colors: Dictionary of color hex to index for Orca mode.
-    :param next_resource_id: Next available resource ID.
-    :param export_pbr: Whether to export PBR display properties (default True).
-    :return: Tuple of (name_to_index mapping, updated next_resource_id, material_resource_id, basematerials_element).
-    """
-    name_to_index = {}
-    next_index = 0
-    material_resource_id = "-1"
-
-    # Create an element lazily. We don't want to create an element if there are no materials to write.
-    basematerials_element = None
-
-    # Orca Slicer mode: Use Materials extension m:colorgroup (vendor-specific)
-    if use_orca_format and vertex_colors:
-        # Sort colors by their index to maintain consistent ordering
-        sorted_colors = sorted(vertex_colors.items(), key=lambda x: x[1])
-
-        # Create a colorgroup for each color (Orca expects one color per group)
-        for color_hex, color_index in sorted_colors:
-            colorgroup_id = next_resource_id
-            next_resource_id += 1
-
-            # Store the first colorgroup ID as our material resource ID
-            if color_index == 0:
-                material_resource_id = str(colorgroup_id)
-
-            # Create m:colorgroup element
-            colorgroup_element = xml.etree.ElementTree.SubElement(
-                resources_element,
-                f"{{{MATERIAL_NAMESPACE}}}colorgroup",
-                attrib={"id": str(colorgroup_id)},
-            )
-            # Add m:color child with the color
-            xml.etree.ElementTree.SubElement(
-                colorgroup_element,
-                f"{{{MATERIAL_NAMESPACE}}}color",
-                attrib={"color": color_hex},
-            )
-            # Map color hex to colorgroup ID for object/triangle assignment
-            name_to_index[color_hex] = colorgroup_id
-
-        log.info(f"Created {len(sorted_colors)} colorgroups for Orca: {name_to_index}")
-        return name_to_index, next_resource_id, material_resource_id, None
-
-    # Collect PBR data for all materials first
-    pbr_materials = []  # List of (material_name, pbr_data_dict)
-
-    # Normal material handling (when not in Orca mode)
-    for blender_object in blender_objects:
-        for material_slot in blender_object.material_slots:
-            material = material_slot.material
-
-            # Skip empty material slots
-            if material is None:
-                continue
-
-            # Cache material name to protect Unicode characters from garbage collection
-            material_name = str(material.name)
-            if material_name in name_to_index:
-                continue
-
-            # Wrap this material into a principled render node, to convert its color to sRGB.
-            principled = bpy_extras.node_shader_utils.PrincipledBSDFWrapper(
-                material, is_readonly=True
-            )
-            color = principled.base_color
-            red = min(255, round(color[0] * 255))
-            green = min(255, round(color[1] * 255))
-            blue = min(255, round(color[2] * 255))
-            alpha = principled.alpha
-            if alpha >= 1.0:
-                color_hex = "#%0.2X%0.2X%0.2X" % (red, green, blue)
-            else:
-                alpha = min(255, round(alpha * 255))
-                color_hex = "#%0.2X%0.2X%0.2X%0.2X" % (red, green, blue, alpha)
-
-            if basematerials_element is None:
-                material_resource_id = str(next_resource_id)
-                next_resource_id += 1
-                basematerials_element = xml.etree.ElementTree.SubElement(
-                    resources_element,
-                    f"{{{MODEL_NAMESPACE}}}basematerials",
-                    attrib={f"{{{MODEL_NAMESPACE}}}id": material_resource_id},
-                )
-            xml.etree.ElementTree.SubElement(
-                basematerials_element,
-                f"{{{MODEL_NAMESPACE}}}base",
-                attrib={
-                    f"{{{MODEL_NAMESPACE}}}name": material_name,
-                    f"{{{MODEL_NAMESPACE}}}displaycolor": color_hex,
-                },
-            )
-            name_to_index[material_name] = next_index
-
-            # Extract PBR properties for this material
-            if export_pbr:
-                pbr_data = _extract_pbr_from_material(material, principled)
-                pbr_materials.append((material_name, pbr_data))
-
-            next_index += 1
-
-    # Write PBR display properties if we have any meaningful PBR data
-    if export_pbr and pbr_materials and basematerials_element is not None:
-        next_resource_id = _write_pbr_display_properties(
-            resources_element, basematerials_element, material_resource_id,
-            pbr_materials, next_resource_id
-        )
-
-    return name_to_index, next_resource_id, material_resource_id, basematerials_element
-
-
-def _extract_pbr_from_material(material: bpy.types.Material,
-                               principled: bpy_extras.node_shader_utils.PrincipledBSDFWrapper) -> Dict:
-    """
-    Extract PBR properties from a Blender material's Principled BSDF.
-
-    Supports all three 3MF Materials Extension workflows:
-    - Metallic: metallicness, roughness
-    - Specular: specularcolor (RGB), glossiness
-    - Translucent: refractiveindex (RGB), roughness, attenuation (RGB)
-
-    :param material: The Blender material
-    :param principled: PrincipledBSDFWrapper for the material
-    :return: Dictionary with PBR properties for export
-    """
-    # Helper to safely convert values to float (handles MagicMock in unit tests)
-    def safe_float(value, default):
-        try:
-            return float(value)
-        except (TypeError, ValueError):
-            return default
-
-    def safe_color(value, default):
-        """Safely extract RGB tuple from color value."""
-        try:
-            if hasattr(value, '__iter__') and len(value) >= 3:
-                return (float(value[0]), float(value[1]), float(value[2]))
-            return default
-        except (TypeError, ValueError, IndexError):
-            return default
-
-    # Core PBR properties
-    metallic = safe_float(principled.metallic, 0.0)
-    roughness = safe_float(principled.roughness, 0.5)
-    specular_ior_level = safe_float(principled.specular, 0.5)
-    ior = safe_float(principled.ior, 1.45)
-    base_color = safe_color(principled.base_color, (0.8, 0.8, 0.8))
-
-    # Calculate specular color for specular workflow
-    # In Blender 4.0+, Specular IOR Level controls Fresnel reflectance
-    # Default 0.5 = 4% reflectance at normal incidence (#383838 in sRGB)
-    # Scale specular color by the specular IOR level
-    default_specular_gray = 0.22  # ~4% reflectance in linear space (sRGB #383838)
-    specular_intensity = specular_ior_level * 2 * default_specular_gray  # Scale from 0-1 to color
-
-    # Check for specular tint (tints specular with base color)
-    # In Blender 4.x, Specular Tint is an RGBA color where WHITE (1,1,1) = no tint
-    # We need to detect if it's NOT white to know if tinting is applied
-    specular_tint_color = None
-    try:
-        if material.node_tree:
-            for node in material.node_tree.nodes:
-                if node.type == 'BSDF_PRINCIPLED':
-                    if 'Specular Tint' in node.inputs:
-                        tint_input = node.inputs['Specular Tint']
-                        if hasattr(tint_input, 'default_value'):
-                            val = tint_input.default_value
-                            # Blender 4.x: It's an RGBA color
-                            if hasattr(val, '__iter__') and len(val) >= 3:
-                                # Check if it's NOT white (default = no tint)
-                                r, g, b = float(val[0]), float(val[1]), float(val[2])
-                                if abs(r - 1.0) > 0.01 or abs(g - 1.0) > 0.01 or abs(b - 1.0) > 0.01:
-                                    specular_tint_color = (r, g, b)
-                            else:
-                                # Old Blender: It's a float 0-1
-                                tint_amount = safe_float(val, 0.0)
-                                if tint_amount > 0.01:
-                                    specular_tint_color = base_color[:3]
-                    break
-    except (TypeError, AttributeError):
-        pass
-
-    # Calculate final specular color
-    if specular_tint_color is not None:
-        # Use the tint color directly (it already incorporates the tint)
-        specular_color = specular_tint_color
-    else:
-        # Neutral gray specular based on specular IOR level
-        specular_color = (specular_intensity, specular_intensity, specular_intensity)
-
-    pbr_data = {
-        "metallic": metallic,
-        "roughness": roughness,
-        "glossiness": 1.0 - roughness,  # Specular workflow uses glossiness (inverse)
-        "specular_ior_level": specular_ior_level,
-        "specular_color": specular_color,
-        "ior": ior,
-        "transmission": 0.0,
-        "attenuation": None,
-    }
-
-    # Get transmission from node tree (PrincipledBSDFWrapper may not expose it directly)
-    try:
-        if material.node_tree:
-            for node in material.node_tree.nodes:
-                if node.type == 'BSDF_PRINCIPLED':
-                    # Blender 4.0+ uses 'Transmission Weight' instead of 'Transmission'
-                    if 'Transmission Weight' in node.inputs:
-                        pbr_data["transmission"] = safe_float(
-                            node.inputs['Transmission Weight'].default_value, 0.0)
-                    elif 'Transmission' in node.inputs:
-                        pbr_data["transmission"] = safe_float(
-                            node.inputs['Transmission'].default_value, 0.0)
-                    break
-    except (TypeError, AttributeError):
-        # Handle mocked objects in unit tests
-        pass
-
-    # Check for stored 3MF attenuation from round-trip
-    try:
-        if "3mf_attenuation" in material:
-            pbr_data["attenuation"] = tuple(material["3mf_attenuation"])
-    except TypeError:
-        # Handle mocked objects in unit tests
-        pass
-
-    # Check for stored 3MF specular_color from round-trip (preserves exact color)
-    try:
-        if "3mf_specular_color" in material:
-            pbr_data["specular_color"] = tuple(material["3mf_specular_color"])
-    except TypeError:
-        # Handle mocked objects in unit tests
-        pass
-
-    # Check for stored 3MF transmission from round-trip
-    try:
-        if "3mf_transmission" in material:
-            pbr_data["transmission"] = float(material["3mf_transmission"])
-    except TypeError:
-        # Handle mocked objects in unit tests
-        pass
-
-    return pbr_data
-
-
-def _write_pbr_display_properties(resources_element: xml.etree.ElementTree.Element,
-                                  basematerials_element: xml.etree.ElementTree.Element,
-                                  basematerials_id: str,
-                                  pbr_materials: List[Tuple[str, Dict]],
-                                  next_resource_id: int) -> int:
-    """
-    Write PBR display properties for materials.
-
-    Supports all three 3MF Materials Extension workflows:
-    - Translucent (transmission > 0.01): translucentdisplayproperties
-    - Metallic (metallic > 0.5): pbmetallicdisplayproperties
-    - Specular (default for dielectrics): pbspeculardisplayproperties
-
-    Per-material workflow selection based on material characteristics.
-
-    :param resources_element: The <resources> element to add display properties to
-    :param basematerials_element: The <basematerials> element to link
-    :param basematerials_id: The ID of the basematerials element
-    :param pbr_materials: List of (name, pbr_data) tuples
-    :param next_resource_id: Next available resource ID
-    :return: Updated next_resource_id
-    """
-    if not pbr_materials:
-        return next_resource_id
-
-    # Check if any material has meaningful PBR data (not just defaults)
-    def has_meaningful_pbr(pbr):
-        metallic_check = pbr.get("metallic", 0) > 0.01
-        roughness_check = pbr.get("roughness", 1) < 0.99
-        transmission_check = pbr.get("transmission", 0) > 0.01
-        specular_check = pbr.get("specular_ior_level", 0.5) != 0.5
-        return metallic_check or roughness_check or transmission_check or specular_check
-    has_pbr_data = any(has_meaningful_pbr(pbr) for _, pbr in pbr_materials)
-
-    if not has_pbr_data:
-        log.debug("No meaningful PBR data to export, skipping display properties")
-        return next_resource_id
-
-    # Categorize materials by workflow
-    # Any material with metallic > 0.01 uses metallic workflow (not just > 0.5!)
-    # Any material with transmission > 0.01 uses translucent workflow
-    translucent_materials = []
-    metallic_materials = []
-    specular_materials = []
-
-    for material_name, pbr in pbr_materials:
-        transmission = pbr.get("transmission", 0)
-        metallic = pbr.get("metallic", 0)
-
-        if transmission > 0.01:
-            translucent_materials.append((material_name, pbr))
-        elif metallic > 0.01:  # Fixed: was > 0.5, which lost partial metallic values!
-            metallic_materials.append((material_name, pbr))
-        else:
-            specular_materials.append((material_name, pbr))
-
-    # Write each workflow type as needed
-    # Note: 3MF allows only ONE displaypropertiesid per basematerials,
-    # so we choose the dominant workflow for all materials
-    # Priority: translucent > metallic > specular (most to least specialized)
-
-    if translucent_materials:
-        # Write translucentdisplayproperties for ALL materials
-        display_props_id = str(next_resource_id)
-        next_resource_id += 1
-
-        translucent_props = xml.etree.ElementTree.SubElement(
-            resources_element,
-            f"{{{MATERIAL_NAMESPACE}}}translucentdisplayproperties",
-            attrib={"id": display_props_id},
-        )
-
-        for material_name, pbr in pbr_materials:
-            ior = pbr.get("ior", 1.45)
-            roughness = pbr.get("roughness", 0.0)
-            attenuation = pbr.get("attenuation")
-            transmission = pbr.get("transmission", 1.0)
-
-            attrib = {
-                "name": material_name,
-                "refractiveindex": f"{ior:.6g} {ior:.6g} {ior:.6g}",
-                "roughness": f"{roughness:.6g}",
-            }
-
-            # Store transmission as custom attribute for round-trip
-            # (3MF spec doesn't have transmission in translucent - it's assumed 1.0)
-            if transmission < 0.99:
-                attrib["blender_transmission"] = f"{transmission:.6g}"
-
-            if attenuation:
-                attrib["attenuation"] = f"{attenuation[0]:.6g} {attenuation[1]:.6g} {attenuation[2]:.6g}"
-            else:
-                attrib["attenuation"] = "0 0 0"
-
-            xml.etree.ElementTree.SubElement(
-                translucent_props,
-                f"{{{MATERIAL_NAMESPACE}}}translucent",
-                attrib=attrib,
-            )
-
-        basematerials_element.set("displaypropertiesid", display_props_id)
-        log.info(f"Exported {len(pbr_materials)} translucent display properties (ID: {display_props_id})")
-
-    elif metallic_materials:
-        # Write pbmetallicdisplayproperties for ALL materials
-        display_props_id = str(next_resource_id)
-        next_resource_id += 1
-
-        metallic_props = xml.etree.ElementTree.SubElement(
-            resources_element,
-            f"{{{MATERIAL_NAMESPACE}}}pbmetallicdisplayproperties",
-            attrib={"id": display_props_id},
-        )
-
-        for material_name, pbr in pbr_materials:
-            metallic = pbr.get("metallic", 0.0)
-            roughness = pbr.get("roughness", 1.0)
-
-            xml.etree.ElementTree.SubElement(
-                metallic_props,
-                f"{{{MATERIAL_NAMESPACE}}}pbmetallic",
-                attrib={
-                    "name": material_name,
-                    "metallicness": f"{metallic:.6g}",
-                    "roughness": f"{roughness:.6g}",
-                },
-            )
-
-        basematerials_element.set("displaypropertiesid", display_props_id)
-        log.info(f"Exported {len(pbr_materials)} metallic display properties (ID: {display_props_id})")
-
-    else:
-        # Write pbspeculardisplayproperties (dielectric materials)
-        display_props_id = str(next_resource_id)
-        next_resource_id += 1
-
-        specular_props = xml.etree.ElementTree.SubElement(
-            resources_element,
-            f"{{{MATERIAL_NAMESPACE}}}pbspeculardisplayproperties",
-            attrib={"id": display_props_id},
-        )
-
-        for material_name, pbr in pbr_materials:
-            specular_color = pbr.get("specular_color", (0.22, 0.22, 0.22))
-            glossiness = pbr.get("glossiness", 0.5)
-
-            # Convert linear specular color to sRGB hex
-            sr = min(255, max(0, int(specular_color[0] * 255)))
-            sg = min(255, max(0, int(specular_color[1] * 255)))
-            sb = min(255, max(0, int(specular_color[2] * 255)))
-            specular_hex = "#%02X%02X%02X" % (sr, sg, sb)
-
-            xml.etree.ElementTree.SubElement(
-                specular_props,
-                f"{{{MATERIAL_NAMESPACE}}}pbspecular",
-                attrib={
-                    "name": material_name,
-                    "specularcolor": specular_hex,
-                    "glossiness": f"{glossiness:.6g}",
-                },
-            )
-
-        basematerials_element.set("displaypropertiesid", display_props_id)
-        log.info(f"Exported {len(pbr_materials)} specular display properties (ID: {display_props_id})")
-
-    return next_resource_id
-
-
-# =============================================================================
-# Texture Export (3MF Materials Extension)
-# =============================================================================
-
-def detect_textured_materials(blender_objects: List[bpy.types.Object]) -> Dict[str, Dict]:
-    """
-    Detect materials with Image Texture nodes connected to Base Color.
-
-    Returns a dictionary of material name -> texture info:
-    {
-        "material_name": {
-            "image": bpy.types.Image,
-            "tilestyleu": str,
-            "tilestylev": str,
-            "filter": str,
-            "original_path": str  # from custom property if imported
-        }
-    }
-    """
-    textured_materials = {}
-
-    for blender_object in blender_objects:
-        for material_slot in blender_object.material_slots:
-            material = material_slot.material
-            if material is None or not material.use_nodes:
-                continue
-
-            material_name = str(material.name)
-            if material_name in textured_materials:
-                continue
-
-            # Find Image Texture node connected to Principled BSDF Base Color
-            image_info = _find_base_color_texture(material)
-            if image_info:
-                textured_materials[material_name] = image_info
-                log.debug(f"Detected textured material: {material_name}")
-
-    return textured_materials
-
-
-def _find_base_color_texture(material: bpy.types.Material) -> Optional[Dict]:
-    """
-    Find Image Texture node connected to Principled BSDF Base Color input.
-
-    :param material: Blender material to analyze
-    :return: Dict with image info, or None if not found
-    """
-    if not material.node_tree:
-        return None
-
-    nodes = material.node_tree.nodes
-    links = material.node_tree.links
-
-    # Find Principled BSDF node
-    principled = None
-    for node in nodes:
-        if node.type == 'BSDF_PRINCIPLED':
-            principled = node
-            break
-
-    if not principled:
-        return None
-
-    # Check Base Color input for image texture
-    base_color_input = principled.inputs.get('Base Color')
-    if not base_color_input or not base_color_input.is_linked:
-        return None
-
-    # Trace back to find Image Texture node
-    for link in links:
-        if link.to_socket == base_color_input:
-            from_node = link.from_node
-            if from_node.type == 'TEX_IMAGE' and from_node.image:
-                image = from_node.image
-
-                # Determine tile style from extension mode
-                extension = getattr(from_node, 'extension', 'REPEAT')
-                if extension == 'CLIP':
-                    tilestyleu = "clamp"
-                    tilestylev = "clamp"
-                elif extension == 'EXTEND':
-                    tilestyleu = "mirror"  # Closest approximation
-                    tilestylev = "mirror"
-                else:
-                    tilestyleu = "wrap"
-                    tilestylev = "wrap"
-
-                # Check for stored metadata from import
-                tilestyleu = material.get("3mf_texture_tilestyleu", tilestyleu)
-                tilestylev = material.get("3mf_texture_tilestylev", tilestylev)
-                filter_mode = material.get("3mf_texture_filter", "auto")
-                original_path = material.get("3mf_texture_path", "")
-
-                # Determine filter from interpolation
-                interpolation = getattr(from_node, 'interpolation', 'Linear')
-                if interpolation == 'Closest':
-                    filter_mode = "nearest"
-                elif filter_mode not in ("linear", "nearest"):
-                    filter_mode = "auto"
-
-                return {
-                    "image": image,
-                    "tilestyleu": tilestyleu,
-                    "tilestylev": tilestylev,
-                    "filter": filter_mode,
-                    "original_path": original_path,
-                }
-
-    return None
-
-
-def _find_texture_from_input(material: bpy.types.Material,
-                             input_name: str,
-                             non_color: bool = False) -> Optional[Dict]:
-    """
-    Find Image Texture node connected to a specific Principled BSDF input.
-
-    :param material: Blender material to analyze
-    :param input_name: Name of the Principled BSDF input (e.g., 'Roughness', 'Metallic')
-    :param non_color: Whether this texture should be non-color data
-    :return: Dict with image info, or None if not found
-    """
-    if not material.node_tree:
-        return None
-
-    nodes = material.node_tree.nodes
-    links = material.node_tree.links
-
-    # Find Principled BSDF node
-    principled = None
-    for node in nodes:
-        if node.type == 'BSDF_PRINCIPLED':
-            principled = node
-            break
-
-    if not principled:
-        return None
-
-    # Check the specified input for image texture
-    target_input = principled.inputs.get(input_name)
-    if not target_input or not target_input.is_linked:
-        return None
-
-    # Trace back to find Image Texture node (may go through other nodes)
-    def trace_to_image_texture(socket, depth=0):
-        """Recursively trace back through nodes to find Image Texture."""
-        if depth > 10:  # Prevent infinite loops
-            return None
-        for link in links:
-            if link.to_socket == socket:
-                from_node = link.from_node
-                if from_node.type == 'TEX_IMAGE' and from_node.image:
-                    return from_node
-                # Check if we can trace through this node (e.g., Normal Map, Invert)
-                if from_node.inputs:
-                    # Try common input names
-                    for input_sock in from_node.inputs:
-                        if input_sock.is_linked:
-                            result = trace_to_image_texture(input_sock, depth + 1)
-                            if result:
-                                return result
-        return None
-
-    tex_node = trace_to_image_texture(target_input)
-    if not tex_node:
-        return None
-
-    image = tex_node.image
-
-    # Determine tile style from extension mode
-    extension = getattr(tex_node, 'extension', 'REPEAT')
-    if extension == 'CLIP':
-        tilestyleu = "clamp"
-        tilestylev = "clamp"
-    elif extension == 'EXTEND':
-        tilestyleu = "mirror"  # Closest approximation
-        tilestylev = "mirror"
-    else:
-        tilestyleu = "wrap"
-        tilestylev = "wrap"
-
-    # Determine filter from interpolation
-    interpolation = getattr(tex_node, 'interpolation', 'Linear')
-    if interpolation == 'Closest':
-        filter_mode = "nearest"
-    else:
-        filter_mode = "auto"
-
-    return {
-        "image": image,
-        "tilestyleu": tilestyleu,
-        "tilestylev": tilestylev,
-        "filter": filter_mode,
-        "non_color": non_color,
-    }
-
-
-def detect_pbr_textured_materials(blender_objects: List[bpy.types.Object]) -> Dict[str, Dict]:
-    """
-    Detect materials with PBR texture nodes connected to Principled BSDF.
-
-    Detects textures connected to:
-    - Base Color
-    - Roughness
-    - Metallic
-    - Normal (through Normal Map node)
-
-    Returns a dictionary of material name -> PBR texture info:
-    {
-        "material_name": {
-            "base_color": {image info dict},     # or None
-            "roughness": {image info dict},       # or None
-            "metallic": {image info dict},        # or None
-            "normal": {image info dict},          # or None
-        }
-    }
-    """
-    pbr_materials = {}
-
-    for blender_object in blender_objects:
-        for material_slot in blender_object.material_slots:
-            material = material_slot.material
-            if material is None or not material.use_nodes:
-                continue
-
-            material_name = str(material.name)
-            if material_name in pbr_materials:
-                continue
-
-            # Check for PBR textures
-            base_color = _find_base_color_texture(material)
-            roughness = _find_texture_from_input(material, 'Roughness', non_color=True)
-            metallic = _find_texture_from_input(material, 'Metallic', non_color=True)
-            normal = _find_texture_from_input(material, 'Normal', non_color=True)
-
-            # Only include if at least one texture is found
-            if base_color or roughness or metallic or normal:
-                pbr_materials[material_name] = {
-                    "base_color": base_color,
-                    "roughness": roughness,
-                    "metallic": metallic,
-                    "normal": normal,
-                }
-                texture_types = [t for t in ["base_color", "roughness", "metallic", "normal"]
-                                 if pbr_materials[material_name][t]]
-                log.debug(f"Detected PBR material '{material_name}' with textures: {texture_types}")
-
-    return pbr_materials
-
-
-def write_textures_to_archive(archive: zipfile.ZipFile,
-                              textured_materials: Dict[str, Dict]
-                              ) -> Dict[str, str]:
-    """
-    Write texture images to the 3MF archive.
-
-    :param archive: The 3MF zip archive
-    :param textured_materials: Dict from detect_textured_materials()
-    :return: Dict mapping image name -> archive path (e.g. "/3D/Texture/image.png")
-    """
-    image_to_path = {}
-    texture_folder = "3D/Texture"
-
-    for mat_name, tex_info in textured_materials.items():
-        image = tex_info["image"]
-        image_name = str(image.name)
-
-        # Skip if already written
-        if image_name in image_to_path:
-            continue
-
-        # Determine output format and path
-        # Prefer original format if available, otherwise use PNG
-        original_path = tex_info.get("original_path", "")
-        if original_path and original_path.lower().endswith('.jpg'):
-            ext = ".jpg"
-        elif original_path and original_path.lower().endswith('.jpeg'):
-            ext = ".jpeg"
-        else:
-            ext = ".png"
-
-        # Sanitize filename
-        safe_name = "".join(c if c.isalnum() or c in "._-" else "_" for c in image_name)
-        if not safe_name.lower().endswith(ext):
-            safe_name += ext
-
-        archive_path = f"{texture_folder}/{safe_name}"
-        full_archive_path = f"/{archive_path}"
-
-        try:
-            # Save image to temporary file, then add to archive
-            with tempfile.NamedTemporaryFile(suffix=ext, delete=False) as tmp:
-                tmp_path = tmp.name
-
-            # Determine format for save
-            if ext in (".jpg", ".jpeg"):
-                file_format = 'JPEG'
-            else:
-                file_format = 'PNG'
-
-            # Save image
-            original_filepath = image.filepath_raw
-            image.filepath_raw = tmp_path
-            image.file_format = file_format
-            image.save()
-            image.filepath_raw = original_filepath
-
-            # Add to archive
-            archive.write(tmp_path, archive_path)
-            os.unlink(tmp_path)
-
-            image_to_path[image_name] = full_archive_path
-            log.info(f"Wrote texture '{image_name}' to {archive_path}")
-
-        except Exception as e:
-            log.warning(f"Failed to write texture '{image_name}': {e}")
-            # Try alternative: if image is packed, write from packed data
-            if image.packed_file:
-                try:
-                    archive.writestr(archive_path, image.packed_file.data)
-                    image_to_path[image_name] = full_archive_path
-                    log.info(f"Wrote packed texture '{image_name}' to {archive_path}")
-                except Exception as e2:
-                    log.error(f"Failed to write packed texture '{image_name}': {e2}")
-
-    return image_to_path
-
-
-def write_texture_relationships(archive: zipfile.ZipFile,
-                                image_to_path: Dict[str, str]) -> None:
-    """
-    Write the model's relationship file to declare texture resources.
-
-    According to the 3MF spec and OPC conventions, textures should be declared
-    as relationships in 3D/_rels/3dmodel.model.rels.
-
-    :param archive: The 3MF zip archive
-    :param image_to_path: Dict mapping image name -> archive path
-    """
-    if not image_to_path:
-        return
-
-    # Create relationships XML
-    relationships_element = xml.etree.ElementTree.Element(
-        f"{{{RELS_NAMESPACE}}}Relationships"
-    )
-
-    # Add a relationship for each texture
-    rel_id = 1
-    for image_name, archive_path in image_to_path.items():
-        rel_element = xml.etree.ElementTree.SubElement(
-            relationships_element,
-            f"{{{RELS_NAMESPACE}}}Relationship"
-        )
-        rel_element.attrib["Type"] = TEXTURE_REL
-        rel_element.attrib["Target"] = archive_path
-        rel_element.attrib["Id"] = f"rel{rel_id}"
-        rel_id += 1
-
-    # Write to archive at 3D/_rels/3dmodel.model.rels
-    rels_path = "3D/_rels/3dmodel.model.rels"
-    tree = xml.etree.ElementTree.ElementTree(relationships_element)
-    with archive.open(rels_path, "w") as f:
-        tree.write(
-            f,
-            xml_declaration=True,
-            encoding="UTF-8",
-        )
-
-    log.info(f"Wrote {len(image_to_path)} texture relationships to {rels_path}")
-
-
-def write_texture_resources(resources_element: xml.etree.ElementTree.Element,
-                            textured_materials: Dict[str, Dict],
-                            image_to_path: Dict[str, str],
-                            next_resource_id: int,
-                            precision: int = 6) -> Tuple[Dict[str, int], int]:
-    """
-    Write texture2d and texture2dgroup elements for textured materials.
-
-    :param resources_element: The <resources> element
-    :param textured_materials: Dict from detect_textured_materials()
-    :param image_to_path: Dict from write_textures_to_archive()
-    :param next_resource_id: Next available resource ID
-    :param precision: Decimal places for UV coordinates
-    :return: Tuple of (material_name -> texture_group_id mapping, updated next_resource_id)
-    """
-    material_to_texture_group = {}
-
-    # Track written textures (image path -> texture2d ID)
-    texture_ids = {}
-
-    for mat_name, tex_info in textured_materials.items():
-        image = tex_info["image"]
-        image_name = str(image.name)
-
-        if image_name not in image_to_path:
-            continue
-
-        archive_path = image_to_path[image_name]
-
-        # Write texture2d element if not already written for this image
-        if archive_path not in texture_ids:
-            texture_id = str(next_resource_id)
-            next_resource_id += 1
-
-            # Determine content type
-            if archive_path.lower().endswith(('.jpg', '.jpeg')):
-                contenttype = TEXTURE_MIMETYPE_JPEG
-            else:
-                contenttype = TEXTURE_MIMETYPE_PNG
-
-            texture_attrib = {
-                "id": texture_id,
-                "path": archive_path,
-                "contenttype": contenttype,
-            }
-
-            # Add optional tile style attributes if not default
-            if tex_info["tilestyleu"] != "wrap":
-                texture_attrib["tilestyleu"] = tex_info["tilestyleu"]
-            if tex_info["tilestylev"] != "wrap":
-                texture_attrib["tilestylev"] = tex_info["tilestylev"]
-            if tex_info["filter"] != "auto":
-                texture_attrib["filter"] = tex_info["filter"]
-
-            xml.etree.ElementTree.SubElement(
-                resources_element,
-                f"{{{MATERIAL_NAMESPACE}}}texture2d",
-                attrib=texture_attrib,
-            )
-            texture_ids[archive_path] = texture_id
-            log.debug(f"Created texture2d ID {texture_id} for {archive_path}")
-
-        # Create texture2dgroup for this material
-        # Note: tex2coord elements will be added when writing triangles
-        texture2d_id = texture_ids[archive_path]
-        group_id = str(next_resource_id)
-        next_resource_id += 1
-
-        group_element = xml.etree.ElementTree.SubElement(
-            resources_element,
-            f"{{{MATERIAL_NAMESPACE}}}texture2dgroup",
-            attrib={
-                "id": group_id,
-                "texid": texture2d_id,
-            },
-        )
-
-        # Store for later tex2coord population
-        material_to_texture_group[mat_name] = {
-            "group_id": group_id,
-            "group_element": group_element,
-            "tex2coords": {},  # UV tuple -> index mapping
-            "next_index": 0,
-            "precision": precision,
-        }
-        log.debug(f"Created texture2dgroup ID {group_id} for material {mat_name}")
-
-    return material_to_texture_group, next_resource_id
-
-
-def get_or_create_tex2coord(texture_group_data: Dict, u: float, v: float) -> int:
-    """
-    Get or create a tex2coord index for a UV coordinate.
-
-    :param texture_group_data: Dict with tex2coords mapping and group_element
-    :param u: U texture coordinate
-    :param v: V texture coordinate
-    :return: Index of the tex2coord in the texture2dgroup
-    """
-    precision = texture_group_data.get("precision", 6)
-    # Round UV to precision for deduplication
-    u_rounded = round(u, precision)
-    v_rounded = round(v, precision)
-    uv_key = (u_rounded, v_rounded)
-
-    if uv_key in texture_group_data["tex2coords"]:
-        return texture_group_data["tex2coords"][uv_key]
-
-    # Create new tex2coord element
-    index = texture_group_data["next_index"]
-    texture_group_data["next_index"] = index + 1
-
-    xml.etree.ElementTree.SubElement(
-        texture_group_data["group_element"],
-        f"{{{MATERIAL_NAMESPACE}}}tex2coord",
-        attrib={
-            "u": f"{u_rounded:.{precision}g}",
-            "v": f"{v_rounded:.{precision}g}",
-        },
-    )
-
-    texture_group_data["tex2coords"][uv_key] = index
-    return index
-
-
-def write_pbr_textures_to_archive(archive: zipfile.ZipFile,
-                                  pbr_materials: Dict[str, Dict]
-                                  ) -> Dict[str, str]:
-    """
-    Write ALL PBR texture images (base_color, roughness, metallic, normal) to the 3MF archive.
-
-    For materials with PBR textures (roughness/metallic), all textures including base color
-    should go through pbmetallictexturedisplayproperties, not texture2dgroup.
-
-    :param archive: The 3MF zip archive
-    :param pbr_materials: Dict from detect_pbr_textured_materials()
-    :return: Dict mapping image name -> archive path
-    """
-    image_to_path = {}
-    texture_folder = "3D/Texture"
-
-    # Collect all unique images from all PBR channels (including base_color)
-    for mat_name, pbr_info in pbr_materials.items():
-        # Only process materials that have roughness or metallic textures
-        # These are the ones that will use pbmetallictexturedisplayproperties
-        if not pbr_info.get('roughness') and not pbr_info.get('metallic'):
-            continue
-
-        # Include base_color along with PBR channels
-        for channel in ['base_color', 'roughness', 'metallic', 'normal']:
-            tex_info = pbr_info.get(channel)
-            if not tex_info:
-                continue
-
-            image = tex_info["image"]
-            image_name = str(image.name)
-
-            if image_name in image_to_path:
-                continue
-
-            # Determine output format - PBR textures typically stay as-is
-            ext = ".png"
-            if image.filepath_raw:
-                if image.filepath_raw.lower().endswith(('.jpg', '.jpeg')):
-                    ext = ".jpg"
-
-            # Sanitize filename
-            safe_name = "".join(c if c.isalnum() or c in "._-" else "_" for c in image_name)
-            if not safe_name.lower().endswith(ext):
-                safe_name += ext
-
-            archive_path = f"{texture_folder}/{safe_name}"
-            full_archive_path = f"/{archive_path}"
-
-            try:
-                # Save image to temporary file, then add to archive
-                with tempfile.NamedTemporaryFile(suffix=ext, delete=False) as tmp:
-                    tmp_path = tmp.name
-
-                file_format = 'JPEG' if ext in (".jpg", ".jpeg") else 'PNG'
-
-                original_filepath = image.filepath_raw
-                image.filepath_raw = tmp_path
-                image.file_format = file_format
-                image.save()
-                image.filepath_raw = original_filepath
-
-                archive.write(tmp_path, archive_path)
-                os.unlink(tmp_path)
-
-                image_to_path[image_name] = full_archive_path
-                log.info(f"Wrote PBR texture '{image_name}' to {archive_path}")
-
-            except Exception as e:
-                log.warning(f"Failed to write PBR texture '{image_name}': {e}")
-                # Try packed data if available
-                if image.packed_file:
-                    try:
-                        archive.writestr(archive_path, image.packed_file.data)
-                        image_to_path[image_name] = full_archive_path
-                        log.info(f"Wrote packed PBR texture '{image_name}' to {archive_path}")
-                    except Exception as e2:
-                        log.error(f"Failed to write packed PBR texture '{image_name}': {e2}")
-
-    return image_to_path
-
-
-def write_pbr_texture_display_properties(
-        resources_element: xml.etree.ElementTree.Element,
-        pbr_materials: Dict[str, Dict],
-        image_to_path: Dict[str, str],
-        next_resource_id: int,
-        basematerials_element: Optional[xml.etree.ElementTree.Element] = None
-) -> Tuple[Dict[str, str], int]:
-    """
-    Write pbmetallictexturedisplayproperties elements for PBR materials.
-
-    Creates texture2d resources for ALL PBR textures (base color, roughness, metallic)
-    and links them via pbmetallictexturedisplayproperties.
-
-    Per 3MF Materials Extension spec, pbmetallictexturedisplayproperties supports:
-    - basecolortextureid: texture for base color (replaces texture2dgroup approach)
-    - metallictextureid: texture for metallic
-    - roughnesstextureid: texture for roughness
-
-    :param resources_element: The <resources> element
-    :param pbr_materials: Dict from detect_pbr_textured_materials()
-    :param image_to_path: Dict mapping image name -> archive path
-    :param next_resource_id: Next available resource ID
-    :param basematerials_element: Optional basematerials element to link via displaypropertiesid
-    :return: Tuple of (material_name -> display_props_id mapping, updated next_resource_id)
-    """
-    material_to_display_props = {}
-    texture_ids = {}  # image path -> texture2d ID
-    first_display_props_id = None  # Track first ID for basematerials linkage
-
-    def get_or_create_texture2d(tex_info: Optional[Dict], tex_type: str) -> str:
-        """Helper to create texture2d resource and return its ID."""
-        nonlocal next_resource_id
-
-        if not tex_info or not tex_info.get('image'):
-            return ""
-
-        image_name = str(tex_info['image'].name)
-        archive_path = image_to_path.get(image_name)
-        if not archive_path:
-            return ""
-
-        if archive_path not in texture_ids:
-            tex_id = str(next_resource_id)
-            next_resource_id += 1
-
-            is_jpeg = archive_path.lower().endswith(('.jpg', '.jpeg'))
-            contenttype = TEXTURE_MIMETYPE_JPEG if is_jpeg else TEXTURE_MIMETYPE_PNG
-
-            xml.etree.ElementTree.SubElement(
-                resources_element,
-                f"{{{MATERIAL_NAMESPACE}}}texture2d",
-                attrib={
-                    "id": tex_id,
-                    "path": archive_path,
-                    "contenttype": contenttype,
-                },
-            )
-            texture_ids[archive_path] = tex_id
-            log.debug(f"Created texture2d ID {tex_id} for {tex_type}: {archive_path}")
-
-        return texture_ids[archive_path]
-
-    for mat_name, pbr_info in pbr_materials.items():
-        base_color_tex = pbr_info.get('base_color')
-        roughness_tex = pbr_info.get('roughness')
-        metallic_tex = pbr_info.get('metallic')
-
-        # Only create pbmetallictexturedisplayproperties if we have PBR textures
-        # (roughness or metallic). Base color alone uses texture2dgroup.
-        if not roughness_tex and not metallic_tex:
-            continue
-
-        # Create texture2d resources for each texture type
-        basecolor_texid = get_or_create_texture2d(base_color_tex, "base_color")
-        roughness_texid = get_or_create_texture2d(roughness_tex, "roughness")
-        metallic_texid = get_or_create_texture2d(metallic_tex, "metallic")
-
-        # Create pbmetallictexturedisplayproperties with all available textures
-        display_props_id = str(next_resource_id)
-        next_resource_id += 1
-
-        if first_display_props_id is None:
-            first_display_props_id = display_props_id
-
-        attrib = {
-            "id": display_props_id,
-            "name": mat_name,
-        }
-        if basecolor_texid:
-            attrib["basecolortextureid"] = basecolor_texid
-        if metallic_texid:
-            attrib["metallictextureid"] = metallic_texid
-        if roughness_texid:
-            attrib["roughnesstextureid"] = roughness_texid
-
-        xml.etree.ElementTree.SubElement(
-            resources_element,
-            f"{{{MATERIAL_NAMESPACE}}}pbmetallictexturedisplayproperties",
-            attrib=attrib,
-        )
-
-        material_to_display_props[mat_name] = display_props_id
-        log.info(
-            f"Created pbmetallictexturedisplayproperties ID {display_props_id} for '{mat_name}' "
-            f"(basecolor={basecolor_texid or 'none'}, roughness={roughness_texid or 'none'}, "
-            f"metallic={metallic_texid or 'none'})"
-        )
-
-    # Link basematerials to display properties
-    # Note: 3MF spec allows only ONE displaypropertiesid per basematerials
-    # Textured PBR takes priority over scalar PBR
-    if first_display_props_id and basematerials_element is not None:
-        basematerials_element.set("displaypropertiesid", first_display_props_id)
-        log.info(f"Linked basematerials to pbmetallictexturedisplayproperties ID {first_display_props_id}")
-
-    return material_to_display_props, next_resource_id
-
-
-def write_prusa_filament_colors(archive: zipfile.ZipFile, vertex_colors: Dict[str, int]) -> None:
-    """
-    Write filament color mapping for PrusaSlicer MMU export.
-
-    Stores colors in Metadata/blender_filament_colors.txt for round-trip import.
-    Format: paint_code=hex_color (one per line)
-    """
-    if not vertex_colors:
-        return
-
-    try:
-        # Sort by colorgroup index to maintain order
-        sorted_colors = sorted(vertex_colors.items(), key=lambda x: x[1])
-
-        # Build color map: colorgroup_id -> hex_color
-        color_lines = []
-        for hex_color, colorgroup_id in sorted_colors:
-            # Map colorgroup_id to paint code
-            if colorgroup_id < len(ORCA_FILAMENT_CODES):
-                paint_code = ORCA_FILAMENT_CODES[colorgroup_id]
-                if paint_code:  # Skip empty (default)
-                    color_lines.append(f"{paint_code}={hex_color}")
-
-        if color_lines:
-            color_data = "\n".join(color_lines)
-            with archive.open("Metadata/blender_filament_colors.txt", "w") as f:
-                f.write(color_data.encode('UTF-8'))
-            log.info(f"Wrote {len(color_lines)} filament colors to metadata")
-    except Exception as e:
-        log.warning(f"Failed to write filament colors: {e}")
-
-
-# =============================================================================
-# Passthrough Materials - Round-trip Support
-# =============================================================================
-
-def write_passthrough_materials(resources_element: xml.etree.ElementTree.Element,
-                                next_resource_id: int) -> Tuple[int, bool]:
-    """
-    Write stored passthrough material data from scene custom properties.
-
-    This writes back compositematerials, multiproperties, textured PBR
-    display properties, colorgroups, and non-textured PBR display properties
-    that were imported but not visually interpreted.
-
-    IDs are remapped to avoid conflicts with newly created materials.
-
-    :param resources_element: The <resources> element
-    :param next_resource_id: Next available resource ID
-    :return: Tuple of (updated next_resource_id, whether any passthrough data was written)
-    """
-    scene = bpy.context.scene
-    any_written = False
-
-    # Check if any passthrough data exists
-    has_composites = bool(scene.get("3mf_compositematerials"))
-    has_multiprops = bool(scene.get("3mf_multiproperties"))
-    has_pbr_tex = bool(scene.get("3mf_pbr_texture_displays"))
-    has_colorgroups = bool(scene.get("3mf_colorgroups"))
-    has_pbr_display = bool(scene.get("3mf_pbr_display_props"))
-    has_textures = bool(scene.get("3mf_textures"))
-    has_tex_groups = bool(scene.get("3mf_texture_groups"))
-
-    if (has_composites or has_multiprops or has_pbr_tex or has_colorgroups
-            or has_pbr_display or has_textures or has_tex_groups):
-        any_written = True
-    else:
-        return next_resource_id, False
-
-    # Build ID remap table: original_id -> new_id
-    # This prevents conflicts with newly created materials
-    # Only remap IDs that would conflict with IDs < next_resource_id
-    id_remap = {}
-
-    # Collect all original IDs that need remapping
-    original_ids = set()
-
-    if has_textures:
-        try:
-            tex_data = json.loads(scene.get("3mf_textures", "{}"))
-            original_ids.update(tex_data.keys())
-        except json.JSONDecodeError:
-            pass
-
-    if has_tex_groups:
-        try:
-            group_data = json.loads(scene.get("3mf_texture_groups", "{}"))
-            original_ids.update(group_data.keys())
-        except json.JSONDecodeError:
-            pass
-
-    if has_colorgroups:
-        try:
-            cg_data = json.loads(scene.get("3mf_colorgroups", "{}"))
-            original_ids.update(cg_data.keys())
-        except json.JSONDecodeError:
-            pass
-
-    if has_pbr_display:
-        try:
-            pbr_data = json.loads(scene.get("3mf_pbr_display_props", "{}"))
-            original_ids.update(pbr_data.keys())
-        except json.JSONDecodeError:
-            pass
-
-    if has_composites:
-        try:
-            comp_data = json.loads(scene.get("3mf_compositematerials", "{}"))
-            original_ids.update(comp_data.keys())
-        except json.JSONDecodeError:
-            pass
-
-    if has_multiprops:
-        try:
-            mp_data = json.loads(scene.get("3mf_multiproperties", "{}"))
-            original_ids.update(mp_data.keys())
-        except json.JSONDecodeError:
-            pass
-
-    if has_pbr_tex:
-        try:
-            pbr_tex_data = json.loads(scene.get("3mf_pbr_texture_displays", "{}"))
-            original_ids.update(pbr_tex_data.keys())
-        except json.JSONDecodeError:
-            pass
-
-    # Find IDs that would conflict with newly created materials (IDs 1 to next_resource_id-1)
-    conflicting_ids = set()
-    for orig_id in original_ids:
-        try:
-            id_int = int(orig_id)
-            if id_int < next_resource_id:
-                conflicting_ids.add(orig_id)
-        except ValueError:
-            pass
-
-    # Only remap conflicting IDs, assign them new unique IDs starting from next_resource_id
-    if conflicting_ids:
-        for orig_id in sorted(conflicting_ids, key=lambda x: int(x) if x.isdigit() else 0):
-            id_remap[orig_id] = str(next_resource_id)
-            next_resource_id += 1
-        log.info(f"Remapped {len(conflicting_ids)} conflicting passthrough IDs: {id_remap}")
-
-    # Update next_resource_id to account for non-conflicting original IDs
-    # This ensures objects don't use IDs that overlap with passthrough
-    max_original_id = max((int(x) for x in original_ids if x.isdigit()), default=0)
-    if max_original_id >= next_resource_id:
-        next_resource_id = max_original_id + 1
-
-    # Write textures first (they may be referenced by other elements)
-    _write_passthrough_textures(resources_element, scene, id_remap)
-
-    # Write texture groups (referenced by multiproperties)
-    _write_passthrough_texture_groups(resources_element, scene, id_remap)
-
-    # Write colorgroups
-    _write_passthrough_colorgroups(resources_element, scene, id_remap)
-
-    # Write non-textured PBR display properties
-    _write_passthrough_pbr_display(resources_element, scene, id_remap)
-
-    # Write compositematerials
-    _write_passthrough_composites(resources_element, scene, id_remap)
-
-    # Write multiproperties
-    _write_passthrough_multiproperties(resources_element, scene, id_remap)
-
-    # Write textured PBR display properties
-    _write_passthrough_pbr_textures(resources_element, scene, id_remap)
-
-    return next_resource_id, any_written
-
-
-def _write_passthrough_composites(resources_element: xml.etree.ElementTree.Element,
-                                  scene: bpy.types.Scene,
-                                  id_remap: Dict[str, str]) -> None:
-    """
-    Write stored compositematerials to XML.
-
-    :param resources_element: The <resources> element
-    :param scene: Blender scene with stored data
-    :param id_remap: Mapping from original IDs to new IDs
-    """
-    stored_data = scene.get("3mf_compositematerials")
-    if not stored_data:
-        return
-
-    try:
-        composite_data = json.loads(stored_data)
-    except json.JSONDecodeError:
-        log.warning("Failed to parse stored compositematerials data")
-        return
-
-    for res_id, comp in composite_data.items():
-        new_id = id_remap.get(res_id, res_id)
-        attrib = {
-            "id": new_id,
-            "matid": id_remap.get(comp["matid"], comp["matid"]),
-            "matindices": comp["matindices"],
-        }
-        if comp.get("displaypropertiesid"):
-            attrib["displaypropertiesid"] = id_remap.get(comp["displaypropertiesid"], comp["displaypropertiesid"])
-
-        comp_element = xml.etree.ElementTree.SubElement(
-            resources_element,
-            f"{{{MATERIAL_NAMESPACE}}}compositematerials",
-            attrib=attrib,
-        )
-
-        # Write composite children
-        for c in comp.get("composites", []):
-            xml.etree.ElementTree.SubElement(
-                comp_element,
-                f"{{{MATERIAL_NAMESPACE}}}composite",
-                attrib={"values": c.get("values", "")},
-            )
-
-        log.debug(f"Wrote passthrough compositematerials {res_id} -> {new_id}")
-
-    log.info(f"Wrote {len(composite_data)} passthrough compositematerials")
-
-
-def _write_passthrough_textures(resources_element: xml.etree.ElementTree.Element,
-                                scene: bpy.types.Scene,
-                                id_remap: Dict[str, str]) -> None:
-    """
-    Write stored texture2d elements to XML.
-
-    :param resources_element: The <resources> element
-    :param scene: Blender scene with stored data
-    :param id_remap: Mapping from original IDs to new IDs
-    """
-    stored_data = scene.get("3mf_textures")
-    if not stored_data:
-        return
-
-    try:
-        texture_data = json.loads(stored_data)
-    except json.JSONDecodeError:
-        log.warning("Failed to parse stored textures data")
-        return
-
-    for res_id, tex in texture_data.items():
-        new_id = id_remap.get(res_id, res_id)
-        attrib = {
-            "id": new_id,
-            "path": tex.get("path", ""),
-            "contenttype": tex.get("contenttype", "image/png"),
-        }
-        # Add optional attributes if not default
-        if tex.get("tilestyleu") and tex.get("tilestyleu") != "wrap":
-            attrib["tilestyleu"] = tex["tilestyleu"]
-        if tex.get("tilestylev") and tex.get("tilestylev") != "wrap":
-            attrib["tilestylev"] = tex["tilestylev"]
-        if tex.get("filter") and tex.get("filter") != "auto":
-            attrib["filter"] = tex["filter"]
-
-        xml.etree.ElementTree.SubElement(
-            resources_element,
-            f"{{{MATERIAL_NAMESPACE}}}texture2d",
-            attrib=attrib,
-        )
-
-        log.debug(f"Wrote passthrough texture2d {res_id} -> {new_id}")
-
-    log.info(f"Wrote {len(texture_data)} passthrough textures")
-
-
-def _write_passthrough_texture_groups(resources_element: xml.etree.ElementTree.Element,
-                                      scene: bpy.types.Scene,
-                                      id_remap: Dict[str, str]) -> None:
-    """
-    Write stored texture2dgroup elements to XML.
-
-    :param resources_element: The <resources> element
-    :param scene: Blender scene with stored data
-    :param id_remap: Mapping from original IDs to new IDs
-    """
-    stored_data = scene.get("3mf_texture_groups")
-    if not stored_data:
-        return
-
-    try:
-        texgroup_data = json.loads(stored_data)
-    except json.JSONDecodeError:
-        log.warning("Failed to parse stored texture groups data")
-        return
-
-    for res_id, tg in texgroup_data.items():
-        new_id = id_remap.get(res_id, res_id)
-        texid = tg.get("texid", "")
-        attrib = {
-            "id": new_id,
-            "texid": id_remap.get(texid, texid),
-        }
-        if tg.get("displaypropertiesid"):
-            dp_id = tg["displaypropertiesid"]
-            attrib["displaypropertiesid"] = id_remap.get(dp_id, dp_id)
-
-        group_element = xml.etree.ElementTree.SubElement(
-            resources_element,
-            f"{{{MATERIAL_NAMESPACE}}}texture2dgroup",
-            attrib=attrib,
-        )
-
-        # Write tex2coord children
-        for coord in tg.get("tex2coords", []):
-            if isinstance(coord, (list, tuple)) and len(coord) >= 2:
-                xml.etree.ElementTree.SubElement(
-                    group_element,
-                    f"{{{MATERIAL_NAMESPACE}}}tex2coord",
-                    attrib={"u": str(coord[0]), "v": str(coord[1])},
-                )
-
-        log.debug(f"Wrote passthrough texture2dgroup {res_id} -> {new_id}")
-
-    log.info(f"Wrote {len(texgroup_data)} passthrough texture groups")
-
-
-def _write_passthrough_colorgroups(resources_element: xml.etree.ElementTree.Element,
-                                   scene: bpy.types.Scene,
-                                   id_remap: Dict[str, str]) -> None:
-    """
-    Write stored colorgroup elements to XML.
-
-    :param resources_element: The <resources> element
-    :param scene: Blender scene with stored data
-    :param id_remap: Mapping from original IDs to new IDs
-    """
-    stored_data = scene.get("3mf_colorgroups")
-    if not stored_data:
-        return
-
-    try:
-        colorgroup_data = json.loads(stored_data)
-    except json.JSONDecodeError:
-        log.warning("Failed to parse stored colorgroups data")
-        return
-
-    for res_id, cg in colorgroup_data.items():
-        new_id = id_remap.get(res_id, res_id)
-        attrib = {"id": new_id}
-        if cg.get("displaypropertiesid"):
-            dp_id = cg["displaypropertiesid"]
-            attrib["displaypropertiesid"] = id_remap.get(dp_id, dp_id)
-
-        group_element = xml.etree.ElementTree.SubElement(
-            resources_element,
-            f"{{{MATERIAL_NAMESPACE}}}colorgroup",
-            attrib=attrib,
-        )
-
-        # Write color children
-        for color in cg.get("colors", []):
-            xml.etree.ElementTree.SubElement(
-                group_element,
-                f"{{{MATERIAL_NAMESPACE}}}color",
-                attrib={"color": color},
-            )
-
-        log.debug(f"Wrote passthrough colorgroup {res_id} -> {new_id}")
-
-    log.info(f"Wrote {len(colorgroup_data)} passthrough colorgroups")
-
-
-def _write_passthrough_pbr_display(resources_element: xml.etree.ElementTree.Element,
-                                   scene: bpy.types.Scene,
-                                   id_remap: Dict[str, str]) -> None:
-    """
-    Write stored non-textured PBR display properties to XML.
-
-    :param resources_element: The <resources> element
-    :param scene: Blender scene with stored data
-    :param id_remap: Mapping from original IDs to new IDs
-    """
-    stored_data = scene.get("3mf_pbr_display_props")
-    if not stored_data:
-        return
-
-    try:
-        pbr_data = json.loads(stored_data)
-    except json.JSONDecodeError:
-        log.warning("Failed to parse stored PBR display props data")
-        return
-
-    for res_id, prop in pbr_data.items():
-        new_id = id_remap.get(res_id, res_id)
-        prop_type = prop.get("type", "metallic")
-        properties = prop.get("properties", [])
-
-        if prop_type == "metallic":
-            element_name = f"{{{MATERIAL_NAMESPACE}}}pbmetallicdisplayproperties"
-            child_name = f"{{{MATERIAL_NAMESPACE}}}pbmetallic"
-        elif prop_type == "specular":
-            element_name = f"{{{MATERIAL_NAMESPACE}}}pbspeculardisplayproperties"
-            child_name = f"{{{MATERIAL_NAMESPACE}}}pbspecular"
-        elif prop_type == "translucent":
-            element_name = f"{{{MATERIAL_NAMESPACE}}}translucentdisplayproperties"
-            child_name = f"{{{MATERIAL_NAMESPACE}}}translucent"
-        else:
-            log.warning(f"Unknown PBR display property type: {prop_type}")
-            continue
-
-        display_element = xml.etree.ElementTree.SubElement(
-            resources_element,
-            element_name,
-            attrib={"id": new_id},
-        )
-
-        # Write child elements with their raw attributes
-        for prop_dict in properties:
-            xml.etree.ElementTree.SubElement(
-                display_element,
-                child_name,
-                attrib=prop_dict,
-            )
-
-        log.debug(f"Wrote passthrough {prop_type} PBR display properties {res_id} -> {new_id}")
-
-    log.info(f"Wrote {len(pbr_data)} passthrough PBR display properties")
-
-
-def _write_passthrough_multiproperties(resources_element: xml.etree.ElementTree.Element,
-                                       scene: bpy.types.Scene,
-                                       id_remap: Dict[str, str]) -> None:
-    """
-    Write stored multiproperties to XML.
-
-    :param resources_element: The <resources> element
-    :param scene: Blender scene with stored data
-    :param id_remap: Mapping from original IDs to new IDs
-    """
-    stored_data = scene.get("3mf_multiproperties")
-    if not stored_data:
-        return
-
-    try:
-        multi_data = json.loads(stored_data)
-    except json.JSONDecodeError:
-        log.warning("Failed to parse stored multiproperties data")
-        return
-
-    for res_id, multi in multi_data.items():
-        new_id = id_remap.get(res_id, res_id)
-        # Remap pids - space-separated list of resource IDs
-        orig_pids = multi["pids"].split()
-        remapped_pids = " ".join(id_remap.get(pid, pid) for pid in orig_pids)
-
-        attrib = {
-            "id": new_id,
-            "pids": remapped_pids,
-        }
-        if multi.get("blendmethods"):
-            attrib["blendmethods"] = multi["blendmethods"]
-
-        multi_element = xml.etree.ElementTree.SubElement(
-            resources_element,
-            f"{{{MATERIAL_NAMESPACE}}}multiproperties",
-            attrib=attrib,
-        )
-
-        # Write multi children
-        for m in multi.get("multis", []):
-            xml.etree.ElementTree.SubElement(
-                multi_element,
-                f"{{{MATERIAL_NAMESPACE}}}multi",
-                attrib={"pindices": m.get("pindices", "")},
-            )
-
-        log.debug(f"Wrote passthrough multiproperties {res_id} -> {new_id}")
-
-    log.info(f"Wrote {len(multi_data)} passthrough multiproperties")
-
-
-def _write_passthrough_pbr_textures(resources_element: xml.etree.ElementTree.Element,
-                                    scene: bpy.types.Scene,
-                                    id_remap: Dict[str, str]) -> None:
-    """
-    Write stored textured PBR display properties to XML.
-
-    :param resources_element: The <resources> element
-    :param scene: Blender scene with stored data
-    :param id_remap: Mapping from original IDs to new IDs
-    """
-    stored_data = scene.get("3mf_pbr_texture_displays")
-    if not stored_data:
-        return
-
-    try:
-        pbr_data = json.loads(stored_data)
-    except json.JSONDecodeError:
-        log.warning("Failed to parse stored PBR texture displays data")
-        return
-
-    for res_id, prop in pbr_data.items():
-        new_id = id_remap.get(res_id, res_id)
-        prop_type = prop.get("type", "specular")
-        factors = prop.get("factors", {})
-
-        if prop_type == "specular":
-            primary_tex = prop.get("primary_texid", "")
-            secondary_tex = prop.get("secondary_texid", "")
-            diffuse_tex = prop.get("basecolor_texid", "")  # diffusetextureid in specular workflow
-            attrib = {
-                "id": new_id,
-                "name": prop.get("name", ""),
-            }
-            # Only include texture IDs if they have values
-            if primary_tex:
-                attrib["speculartextureid"] = id_remap.get(primary_tex, primary_tex)
-            if secondary_tex:
-                attrib["glossinesstextureid"] = id_remap.get(secondary_tex, secondary_tex)
-            if diffuse_tex:
-                attrib["diffusetextureid"] = id_remap.get(diffuse_tex, diffuse_tex)
-            # Add factor attributes
-            for factor_name, factor_value in factors.items():
-                attrib[factor_name] = factor_value
-
-            xml.etree.ElementTree.SubElement(
-                resources_element,
-                f"{{{MATERIAL_NAMESPACE}}}pbspeculartexturedisplayproperties",
-                attrib=attrib,
-            )
-        elif prop_type == "metallic":
-            primary_tex = prop.get("primary_texid", "")
-            secondary_tex = prop.get("secondary_texid", "")
-            basecolor_tex = prop.get("basecolor_texid", "")
-            attrib = {
-                "id": new_id,
-                "name": prop.get("name", ""),
-            }
-            # Only include texture IDs if they have values
-            if primary_tex:
-                attrib["metallictextureid"] = id_remap.get(primary_tex, primary_tex)
-            if secondary_tex:
-                attrib["roughnesstextureid"] = id_remap.get(secondary_tex, secondary_tex)
-            if basecolor_tex:
-                attrib["basecolortextureid"] = id_remap.get(basecolor_tex, basecolor_tex)
-            # Add factor attributes
-            for factor_name, factor_value in factors.items():
-                attrib[factor_name] = factor_value
-
-            xml.etree.ElementTree.SubElement(
-                resources_element,
-                f"{{{MATERIAL_NAMESPACE}}}pbmetallictexturedisplayproperties",
-                attrib=attrib,
-            )
-
-        log.debug(f"Wrote passthrough {prop_type} PBR texture display {res_id} -> {new_id}")
-
-    log.info(f"Wrote {len(pbr_data)} passthrough PBR texture displays")
 
 
 # =============================================================================
@@ -2348,87 +544,6 @@ def write_triangles(
                         if basematerials_resource_id:
                             triangle_element.attrib[pid_name] = str(basematerials_resource_id)
                         triangle_element.attrib[p1_name] = str(material_index)
-
-
-def write_triangle_sets(
-    mesh_element: xml.etree.ElementTree.Element,
-    mesh: bpy.types.Mesh
-) -> None:
-    """
-    Writes triangle sets from Blender mesh attributes into the mesh element.
-
-    :param mesh_element: The <mesh> element of the 3MF document.
-    :param mesh: The Blender mesh containing triangle set attributes to export.
-    """
-    attr_name = "3mf_triangle_set"
-    if attr_name not in mesh.attributes:
-        return
-
-    set_names = mesh.get("3mf_triangle_set_names", [])
-    if not set_names:
-        return
-
-    # Build mapping of set_index -> list of triangle indices
-    num_faces = len(mesh.polygons)
-    set_values = [0] * num_faces
-    mesh.attributes[attr_name].data.foreach_get("value", set_values)
-
-    # Group triangles by set index
-    set_to_triangles: Dict[int, List[int]] = {}
-    for poly_idx, set_idx in enumerate(set_values):
-        if set_idx > 0:
-            if set_idx not in set_to_triangles:
-                set_to_triangles[set_idx] = []
-            set_to_triangles[set_idx].append(poly_idx)
-
-    if not set_to_triangles:
-        return
-
-    trianglesets_element = xml.etree.ElementTree.SubElement(
-        mesh_element, f"{{{TRIANGLE_SETS_NAMESPACE}}}trianglesets"
-    )
-
-    for set_idx, triangle_indices in sorted(set_to_triangles.items()):
-        if set_idx <= len(set_names):
-            set_name = str(set_names[set_idx - 1])
-        else:
-            set_name = f"TriangleSet_{set_idx}"
-
-        triangleset_element = xml.etree.ElementTree.SubElement(
-            trianglesets_element,
-            f"{{{TRIANGLE_SETS_NAMESPACE}}}triangleset"
-        )
-        triangleset_element.attrib["name"] = set_name
-        triangleset_element.attrib["identifier"] = set_name
-
-        triangle_indices = sorted(triangle_indices)
-
-        # Use refrange for consecutive sequences, ref for isolated indices
-        i = 0
-        while i < len(triangle_indices):
-            start = triangle_indices[i]
-            end = start
-            while i + 1 < len(triangle_indices) and triangle_indices[i + 1] == end + 1:
-                i += 1
-                end = triangle_indices[i]
-
-            if end - start >= 2:
-                refrange_element = xml.etree.ElementTree.SubElement(
-                    triangleset_element,
-                    f"{{{TRIANGLE_SETS_NAMESPACE}}}refrange"
-                )
-                refrange_element.attrib["startindex"] = str(start)
-                refrange_element.attrib["endindex"] = str(end)
-            else:
-                for idx in range(start, end + 1):
-                    ref_element = xml.etree.ElementTree.SubElement(
-                        triangleset_element,
-                        f"{{{TRIANGLE_SETS_NAMESPACE}}}ref"
-                    )
-                    ref_element.attrib["index"] = str(idx)
-            i += 1
-
-        log.info(f"Exported triangle set '{set_name}' with {len(triangle_indices)} triangles")
 
 
 # =============================================================================
